api/requirements.txt
Django>=4.0
djangorestframework>=3.15.2
gunicorn>=23.0.0
psycopg2-binary>=2.9.10